Version 0.6.5 released
Changes in the widgets and shortcodes - now you can exclude the Posts
without Featured Image from the Posts list.
Changes in the Fluid Images Layout - now if the Post have no Featured
Image, the block with background displayed instead of the Featured
Image. Background and text color you can choose in the Shortcode
Builder. In widget this feature will be available soon.
Changes in the Thumbnail Layout - now if the Post have no Featured
Image, the block with Thumbnail displayed as a la Drop Cap Layout with
the first letter of the Post title. Background and text color you can
choose in the Shortcode Builder. In widget this feature will be
available soon.
Fixed some bugs in the Shortcode Builder