Version 0.6.9 released
Added the ability to filter the posts by tags (only in a shortcode, in
widgets this ability will be available in the next version). You can
include or exclude post by a specific tag or multiple tags. Now only
work with posts.
Fixed issue with fixed height of a element in mobile version of the
Responsive Grid layout.